+++ azure/storage/blob.go
