+++ azure/storage/client_test.go
