+++ azure/storage/util.go
