+++ azure/virtualmachine/entities.go
@@ -1,8 +1,15 @@
-package vmClient
+package virtualmachine
 
 import (
 	"encoding/xml"
+
+	"github.com/MSOpenTech/azure-sdk-for-go/azure"
 )
+
+//VmClient is used to manage operations on Azure Virtual Machines
+type VirtualMachineClient struct {
+	client azure.Client
+}
 
 type VMDeployment struct {
 	XMLName          xml.Name `xml:"Deployment"`
